requirements.txt
asgiref==3.8.1
awscli==1.35.15
bleach==5.0.1
boto3==1.35.40
botocore==1.35.49
build==1.2.2.post1
click==8.1.7
colorama==0.4.6
cramjam==2.8.3
crispy-tailwind==1.0.3
dj-database-url==2.2.0
Django==5.1.1
django-allauth==64.2.1
django-chartjs==2.3.0
django-crispy-forms==2.3
django-storages==1.14.4
django-summernote==0.8.20.0
dnspython==2.6.1
docutils==0.16
gunicorn==23.0.0
jmespath==1.0.1
pillow==10.4.0
pip-tools==7.4.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pykerberos==1.2.4
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-snappy==0.7.2
rsa==4.7.2
s3transfer==0.10.3
sqlparse==0.5.1
whitenoise==6.7.0
